pyproject.toml
[project]
name = "arxivscraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "numpy>=2.2.5",
    "openai>=1.76.2",
    "pyyaml>=6.0.2",
    "unidecode>=1.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/arXivScraper"]
